fix variable type bug

+++ model_conv.py
@@ -459,7 +459,7 @@ class WGAN_conv(object):
   #get units STA from the discriminator
   def get_critics_output(self, samples):
       #aux = np.load(self.sample_dir+ '/stats_real.npz')
-      output,_,_ = self.Discriminator_sampler(samples)
+      output,_,_ = self.Discriminator_sampler(samples.astype('float32'))
       return output
    
   #this is to save the network parameters
